+++ packages/react-native-push-notification/src/index.ts
@@ -1,2 +1,19 @@
-export * from './model'
-export * from './push-notification-service'
+export {
+  $pushSettings,
+  updateFcmToken,
+  enablePush,
+  disablePush,
+  skipRequestPermissions,
+  resetPushSettings,
+  $pushEnabled,
+  $fcmToken,
+  $needFirstRequestPermission,
+  usePushAvailable,
+  useRegisterDevice,
+  useUnregisterDevice,
+} from './model'
+export {
+  PushNotificationService,
+  usePush,
+  pushService,
+} from './push-notification-service'