+++ packages/react-native-push-notification/CHANGELOG.md
@@ -1,1 +1,31 @@
 # @kode-frontend/react-native-push-notification
+
+## 0.0.4
+
+### Patch Changes
+
+- [#42](https://github.com/appKODE/frontend-depend/pull/42) [`1019a1c`](https://github.com/appKODE/frontend-depend/commit/1019a1c1f4431ea39e29ef64811ceb95afd0c036) Thanks [@DenisAmberCode](https://github.com/DenisAmberCode)! - Fixed build packages
+
+- [#43](https://github.com/appKODE/frontend-depend/pull/43) [`0248955`](https://github.com/appKODE/frontend-depend/commit/024895515a1e8d007d0462e806cb3e6354da3de7) Thanks [@DenisAmberCode](https://github.com/DenisAmberCode)! - Fixed build
+
+## 0.0.3
+
+### Patch Changes
+
+- [#40](https://github.com/appKODE/frontend-depend/pull/40) [`789e71b`](https://github.com/appKODE/frontend-depend/commit/789e71b2260d11d6bccba5c4c4af648f0d9f78f9) Thanks [@DenisAmberCode](https://github.com/DenisAmberCode)! - Fixed build packages
+
+## 0.0.2
+
+### Patch Changes
+
+- [#38](https://github.com/appKODE/frontend-depend/pull/38) [`1edb2b2`](https://github.com/appKODE/frontend-depend/commit/1edb2b25f4140f8cc28bc1110e9eb9ef0006bd38) Thanks [@DenisAmberCode](https://github.com/DenisAmberCode)! - Fixed imports
+
+## 0.0.1
+
+### Patch Changes
+
+- [#32](https://github.com/appKODE/frontend-depend/pull/32) [`8621828`](https://github.com/appKODE/frontend-depend/commit/86218283aca2ff1d2ffab7e65de63d3cca63a009) Thanks [@DenisAmberCode](https://github.com/DenisAmberCode)! - Fixed deps
+
+- [#37](https://github.com/appKODE/frontend-depend/pull/37) [`a9d4ea1`](https://github.com/appKODE/frontend-depend/commit/a9d4ea178730ff74c4feccf9091768a0f3bea88e) Thanks [@DenisAmberCode](https://github.com/DenisAmberCode)! - Added documentation
+
+- [#31](https://github.com/appKODE/frontend-depend/pull/31) [`5dc89c3`](https://github.com/appKODE/frontend-depend/commit/5dc89c35dfd8ad1ae23ac0572feac2fe9f1e80d3) Thanks [@DenisAmberCode](https://github.com/DenisAmberCode)! - Added Push notofication service for React Native apps